requirements.txt
PyQt6>=6.6.0
PyQt6-Charts>=6.6.0
requests>=2.31.0
pandas>=2.1.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-dotenv>=1.0.0
